feat: EPI-802: separate ContactDetails into a sinple component
