+++ packages/web/app/components/RemoveReminderContact.jsx
@@ -5,10 +5,9 @@ import { Box, Divider, Typography } from '@material-ui/core';
 
 import { Colors } from '../constants';
 import { FormConfirmCancelBackRow } from './ButtonRow';
-import { Table } from './Table';
 import { TranslatedText } from './Translation/TranslatedText';
 import { useApi } from '../api';
-import { capitalize } from 'lodash';
+import { ContactDetails } from './ContactDetails';
 
 const StyledHeading = styled(Typography)`
   margin: 6px 0 9px 0;
@@ -23,38 +22,6 @@ const StyledSubHeading = styled(Typography)`
   line-height: 18px;
 `;
 
-const ContactDetailTable = styled(Table)`
-  border: 1px solid ${Colors.outline};
-  box-shadow: none;
-  margin: 22px 0 42px;
-  border-radius: 5px;
-  background: ${Colors.white};
-
-  table {
-    padding-left: 23px;
-    padding-right: 20px;
-
-    thead tr {
-      border-bottom: 1px solid ${Colors.outline};
-
-      th {
-        font-style: normal;
-        background-color: ${Colors.white} !important;
-        padding: 13px 0 12px !important;
-        font-size: 14px;
-        line-height: 18px;
-        font-weight: 500;
-        width: 30%;
-      }
-    }
-
-    tbody tr td {
-      border-bottom: none;
-      padding: 10px 0 16px 1px !important;
-    }
-  }
-`;
-
 const StyledFullWidthContainer = styled(Box)`
   margin: 0 -32px 21px;
   grid-column: 1 / -1;
@@ -64,42 +31,7 @@ const StyledDivider = styled(Divider)`
   border-top: 1px solid ${Colors.outline};
 `;
 
-export const ContactDetails = ({ selectedContact }) => {
-  const columns = [
-    {
-      key: 'name',
-      title: <TranslatedText stringId='patient.details.reminderContacts.field.contact' fallback='Contact' />,
-      sortable: false,
-    },
-    {
-      key: 'relationship.name',
-      title: <TranslatedText stringId='patient.details.reminderContacts.field.relationship' fallback='Relationship' />,
-      sortable: false,
-    },
-    {
-      key: 'method',
-      title: <TranslatedText stringId='patient.details.reminderContacts.field.contactMethod' fallback='Contact method' />,
-      sortable: false,
-      accessor: data => {
-        return data.connectionDetails ? (
-          <TranslatedText
-            stringId={`patient.details.reminderContacts.method.${data.method}`}
-            fallback={capitalize(data.method)}
-          />
-        ) : (
-          <TranslatedText
-            stringId={`patient.details.reminderContacts.method.${data.method}Pending`}
-            fallback={`${capitalize(data.method)} pending`}
-          />
-        );
-      },
-    },
-  ];
-
-  return <ContactDetailTable columns={columns} allowExport={false} data={[selectedContact]} />;
-};
-
-export const RemoveReminderContact = ({ selectedContact, onBack, onClose }) => {
+export const RemoveReminderContact = ({ selectedContact, onBack, onClose, pendingContacts, successContactIds }) => {
   const api = useApi();
 
   const handleDeleteContact = async () => {
@@ -122,7 +54,12 @@ export const RemoveReminderContact = ({ selectedContact, onBack, onClose }) => {
         />
       </StyledSubHeading>
 
-      <ContactDetails selectedContact={selectedContact} />
+      <ContactDetails 
+        selectedContact={selectedContact} 
+        isRemoveModal 
+        pendingContacts={pendingContacts} 
+        successContactIds={successContactIds}  
+      />
 
       <StyledFullWidthContainer>
         <StyledDivider />
